Use long variable in sample usage of onDiscard

+++ sentry-samples/sentry-samples-console/src/main/java/io/sentry/samples/console/Main.java
@@ -8,7 +8,7 @@ import java.util.Collections;
 
 public class Main {
 
-  private static int numberOfDiscardedSpansDueToOverflow = 0;
+  private static long numberOfDiscardedSpansDueToOverflow = 0;
 
   public static void main(String[] args) throws InterruptedException {
     Sentry.init(
